Refactor DockerDataLoader to not be exclusively in tests
