+++ noxfile.py
@@ -45,7 +45,7 @@ def import_data(session: Session) -> None:
     data_dir = _ROOT / "test" / "data"
     sys.path.append(f"{path}")
 
-    from test.integration.conftest import DockerDataLoader
+    from pyexasol_utils.docker_util import DockerDataLoader
 
     loader = DockerDataLoader(
         dsn="127.0.0.1:8563",
